Linter: Fix unused config parameter
SuperSafe: Remove usage of .equals
Scalafix: remove unused imports & private hex function

There is no need to guard against LHS null in Scala equality and isInstanceOf

+++ kamon-core/src/main/scala/kamon/metric/MeasurementUnit.scala
@@ -78,7 +78,7 @@ object MeasurementUnit {
       value
     } else if(from == to)
       value
-    else (from.magnitude.scaleFactor / to.magnitude.scaleFactor) * value.toDouble
+    else (from.magnitude.scaleFactor / to.magnitude.scaleFactor) * value
   }
 
   /**
